johnnyy/VIDSOL-127--refactoring appConfig
refactoring appConfig, moving it to frontend/stores
