--- frontend/src/Context/AppConfig/index.tsx
@@ -1,3 +1,0 @@
-export type { AppConfig } from './AppConfigContext.types';
-export * from './hooks';
-export { default, type AppConfigApi } from './AppConfigContext';
