+++ frontend/src/components/MeetingRoom/VideoDevices/VideoDevices.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect, MouseEvent, ReactElement } from 'react';
 import { Device } from '@vonage/client-sdk-video';
 import { useTranslation } from 'react-i18next';
-import useAppConfig from '@Context/AppConfig/hooks/useAppConfig';
+import { useAppConfig } from '@stores/appConfig';
 import useTheme from '@ui/theme';
 import useDevices from '@hooks/useDevices';
 import usePublisherContext from '@hooks/usePublisherContext';
