+++ frontend/src/stores/appConfig/AppConfigStore.ts
