+++ frontend/src/stores/appConfig/helpers/index.ts
@@ -1,0 +1,2 @@
+export { default as defaultAppConfig } from './defaultAppConfig';
+export { default as mergeAppConfigs } from './mergeAppConfigs';
