+++ frontend/src/stores/appConfig/hooks/index.ts
