+++ frontend/src/stores/appConfig/hooks/useIsCameraControlAllowed.ts
@@ -1,4 +1,4 @@
-import appConfig$ from '../AppConfigContext';
+import appConfig$ from '../AppConfigStore';
 
 const useIsCameraControlAllowed = appConfig$.use.createSelectorHook(
   ({ isAppConfigLoaded, videoSettings }) => isAppConfigLoaded && videoSettings.allowCameraControl
