+++ frontend/src/stores/appConfig/hooks/useShouldShowParticipantList.ts
@@ -1,4 +1,4 @@
-import appConfig$ from '../AppConfigContext';
+import appConfig$ from '../AppConfigStore';
 
 const useShouldShowParticipantList = appConfig$.use.createSelectorHook(
   ({ isAppConfigLoaded, meetingRoomSettings }) =>
