+++ frontend/src/stores/appConfig/hooks/useSuspenseUntilAppConfigReady/index.ts
