+++ frontend/src/stores/appConfig/index.tsx
@@ -1,0 +1,5 @@
+export type { AppConfig } from './types';
+export { default, type AppConfigApi } from './AppConfigStore';
+export * from './hooks';
+export type * from './types';
+export * from './helpers';
