+++ frontend/src/stores/appConfig/types/AudioSettings.ts
@@ -1,0 +1,7 @@
+type AudioSettings = {
+  allowAdvancedNoiseSuppression: boolean;
+  allowAudioOnJoin: boolean;
+  allowMicrophoneControl: boolean;
+};
+
+export default AudioSettings;
