+++ frontend/src/stores/appConfig/types/WaitingRoomSettings.ts
@@ -1,0 +1,5 @@
+type WaitingRoomSettings = {
+  allowDeviceSelection: boolean;
+};
+
+export default WaitingRoomSettings;
