+++ frontend/src/stores/appConfig/types/index.ts
@@ -1,0 +1,5 @@
+export type { default as AppConfig } from './AppConfig';
+export type { default as AudioSettings } from './AudioSettings';
+export type { default as MeetingRoomSettings } from './MeetingRoomSettings';
+export type { default as VideoSettings } from './VideoSettings';
+export type { default as WaitingRoomSettings } from './WaitingRoomSettings';
